+++ EDutils/changelog.md
@@ -1,4 +1,10 @@
 # Changelog
+## 1.2dev
+### Fri  9 Jun 15:08:57 BST 2023
+- added image readers in readers.py
+- ported read_space_group
+- added test for image_readers
+- added smv/img format reader
 
-##Thu  8 Jun 10:38:03 BST 2023
+### Thu  8 Jun 10:38:03 BST 2023
 - added change path functionality